fix(index): useCollection and logout
