+++ src/assets/styles/main.scss
@@ -2,6 +2,7 @@
 @import 'variables';
 @import 'fonts';
 @import 'transitions';
+@import 'scroll';
 
 html {
   font-size: 1vh;
