+++ src/composables/useToken.ts
@@ -3,7 +3,7 @@ import { getJWTToken } from '../helpers/jwt';
 const useToken = async () => {
   const jwtToken = await getJWTToken();
   if (!jwtToken) {
-    navigateTo('/login');
+    await navigateTo('/login');
     return null;
   }
   return jwtToken;
